refactor: Renomeia e reorganiza arquivos
Mais informações:
- Melhora legibilidade e organização do código
